refactor: streamline auth architecture and fix hydration issues
- Replace custom auth logic with Clerk's SignedOut + RedirectToTasks components
- Add proper tRPC prefetching with HydrateClient to prevent hydration mismatches
- Fix tRPC authentication by adding user.getUser to public middleware routes
- Add organization slug validation to prevent cross-org access
- Remove redundant authenticated layout and simplify middleware
- Update TaskChooseOrganization to use correct redirectUrlComplete prop
- Extract user dropdown to separate component for better organization
- Use proper root redirect matchers in both auth and cloud middleware
