+++ apps/cloud/src/app/(trpc)/api/trpc/[trpc]/route.ts
@@ -10,36 +10,37 @@ export const runtime = "nodejs";
  * You should extend this to match your needs
  */
 const setCorsHeaders = (res: Response) => {
-  res.headers.set("Access-Control-Allow-Origin", "*");
-  res.headers.set("Access-Control-Request-Method", "*");
-  res.headers.set("Access-Control-Allow-Methods", "OPTIONS, GET, POST");
-  res.headers.set("Access-Control-Allow-Headers", "*");
+	res.headers.set("Access-Control-Allow-Origin", "*");
+	res.headers.set("Access-Control-Request-Method", "*");
+	res.headers.set("Access-Control-Allow-Methods", "OPTIONS, GET, POST");
+	res.headers.set("Access-Control-Allow-Headers", "*");
 };
 
 export const OPTIONS = () => {
-  const response = new Response(null, {
-    status: 204,
-  });
-  setCorsHeaders(response);
-  return response;
+	const response = new Response(null, {
+		status: 204,
+	});
+	setCorsHeaders(response);
+	return response;
 };
 
 const handler = async (req: NextRequest) => {
-  const response = await fetchRequestHandler({
-    endpoint: "/api/trpc",
-    router: cloudAppRouter,
-    req,
-    createContext: () =>
-      createTRPCContext({
-        headers: req.headers,
-      }),
-    onError({ error, path }) {
-      console.error(`>>> tRPC Error on '${path}'`, error);
-    },
-  });
+	const response = await fetchRequestHandler({
+		endpoint: "/api/trpc",
+		router: cloudAppRouter,
+		req,
+		createContext: () =>
+			createTRPCContext({
+				headers: req.headers,
+			}),
+		onError({ error, path }) {
+			console.error(`>>> tRPC Error on '${path}'`, error);
+		},
+	});
 
-  setCorsHeaders(response);
-  return response;
+	setCorsHeaders(response);
+	return response;
 };
 
 export { handler as GET, handler as POST };
+
