+++ apps/cloud/src/components/app-sidebar.tsx
@@ -2,24 +2,8 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { useClerk } from "@clerk/nextjs";
-import { LayoutDashboard, Settings, LogOut, Key } from "lucide-react";
+import { LayoutDashboard, Settings } from "lucide-react";
 import { Icons } from "@repo/ui/components/icons";
-import { useTRPC } from "~/trpc/react";
-import { useSuspenseQuery } from "@tanstack/react-query";
-import {
-	Avatar,
-	AvatarFallback,
-	AvatarImage,
-} from "@repo/ui/components/ui/avatar";
-import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuLabel,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from "@repo/ui/components/ui/dropdown-menu";
 import {
 	Sidebar,
 	SidebarContent,
@@ -32,8 +16,8 @@ import {
 	SidebarMenuButton,
 	SidebarMenuItem,
 	SidebarTrigger,
-	useSidebar,
 } from "@repo/ui/components/ui/sidebar";
+import { UserDropdownMenu } from "~/components/user-dropdown-menu";
 
 interface NavigationItem {
 	name: string;
@@ -69,112 +53,8 @@ const navigationSections: NavigationSection[] = [
 	},
 ];
 
-function UserSection() {
-	const trpc = useTRPC();
-	const { data: user } = useSuspenseQuery({
-		...trpc.user.getUser.queryOptions(),
-		staleTime: 5 * 60 * 1000, // Cache user data for 5 minutes
-	});
-	const { signOut } = useClerk();
-	const { state } = useSidebar();
 
-	const displayName = 
-		user?.firstName && user?.lastName 
-			? `${user.firstName} ${user.lastName}`
-			: user?.username ?? user?.email ?? "User";
-	const displayEmail = user?.email ?? "";
-	const avatarUrl = user?.imageUrl;
-
-	const handleSignOut = () => {
-		void signOut({ redirectUrl: "/" });
-	};
-
-	return (
-		<SidebarGroup>
-			<SidebarGroupContent>
-				<SidebarMenu>
-					<SidebarMenuItem>
-						<DropdownMenu>
-							<DropdownMenuTrigger asChild>
-								<SidebarMenuButton
-									size="lg"
-									className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
-								>
-									<Avatar className="h-8 w-8">
-										<AvatarImage src={avatarUrl || undefined} alt={displayName} />
-										<AvatarFallback className="bg-primary/10 text-primary">
-											{displayName.charAt(0).toUpperCase()}
-										</AvatarFallback>
-									</Avatar>
-									<div className="grid flex-1 text-left text-sm leading-tight">
-										<span className="truncate font-semibold">{displayName}</span>
-										{displayEmail && state === "expanded" && (
-											<span className="truncate text-xs text-muted-foreground">
-												{displayEmail}
-											</span>
-										)}
-									</div>
-								</SidebarMenuButton>
-							</DropdownMenuTrigger>
-							
-							<DropdownMenuContent 
-								className="w-56" 
-								align="start" 
-								side="right"
-								sideOffset={4}
-							>
-								<DropdownMenuLabel className="font-normal">
-									<div className="flex flex-col space-y-1">
-										<p className="text-sm font-medium leading-none">
-											{displayName}
-										</p>
-										{displayEmail && (
-											<p className="text-xs leading-none text-muted-foreground">
-												{displayEmail}
-											</p>
-										)}
-									</div>
-								</DropdownMenuLabel>
-								
-								<DropdownMenuSeparator />
-								
-								<DropdownMenuItem asChild>
-									<Link href="/settings" className="flex items-center cursor-pointer">
-										<Settings className="mr-2 h-4 w-4" />
-										Settings
-									</Link>
-								</DropdownMenuItem>
-								
-								<DropdownMenuItem asChild>
-									<Link href="/settings/api-keys" className="flex items-center cursor-pointer">
-										<Key className="mr-2 h-4 w-4" />
-										API Keys
-									</Link>
-								</DropdownMenuItem>
-								
-								<DropdownMenuSeparator />
-								
-								<DropdownMenuItem 
-									className="text-red-600 focus:text-red-600 cursor-pointer"
-									onClick={handleSignOut}
-								>
-									<LogOut className="mr-2 h-4 w-4" />
-									Sign Out
-								</DropdownMenuItem>
-							</DropdownMenuContent>
-						</DropdownMenu>
-					</SidebarMenuItem>
-				</SidebarMenu>
-			</SidebarGroupContent>
-		</SidebarGroup>
-	);
-}
-
-interface AppSidebarProps {
-	organizationId?: string;
-}
-
-export function AppSidebar({ organizationId }: AppSidebarProps) {
+export function AppSidebar() {
 	const pathname = usePathname();
 
 	return (
@@ -226,7 +106,7 @@ export function AppSidebar({ organizationId }: AppSidebarProps) {
 			</SidebarContent>
 
 			<SidebarFooter>
-				<UserSection />
+				<UserDropdownMenu />
 			</SidebarFooter>
 		</Sidebar>
 	);
