+++ apps/cloud/src/trpc/react.tsx
@@ -1,10 +1,7 @@
 "use client";
 
-import type {
-	QueryClient} from "@tanstack/react-query";
-import {
-	QueryClientProvider,
-} from "@tanstack/react-query";
+import type { QueryClient } from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { useState } from "react";
 import {
 	createTRPCClient,
